==> ENGE817/Report/1_chunking.ipynb <==
[Chunking](https://community.databricks.com/t5/technical-blog/the-ultimate-guide-to-chunking-strategies-for-rag-applications/ba-p/113089)
- Fixed-Size Chunking (word, char or token counts (with overlaps))
- Semantic Chunking (break at paragraphs or sentences)
- Recursive Chunking
- Adaptive Chunking
- Context-Enriched Chunking
- AI-Driven Dynamic Chunking

%pip install -qU langchain-text-splitters transformers

document = None
with open("./datasets/dsm.md", 'r', encoding='utf-8') as f:
    document = f.read()

## Fixed-Size Chunking

This is the simplest method. This splits based on a given character sequence, which defaults to "\n\n". Chunk length is measured by number of characters.

1. How the text is split: by single character separator.
2. How the chunk size is measured: by number of characters.

To obtain the string content directly, use .split_text.
To create LangChain Document objects (e.g., for use in downstream tasks), use .create_documents.

https://python.langchain.com/docs/how_to/character_text_splitter/

from transformers import AutoTokenizer
from langchain_text_splitters import CharacterTextSplitter

# Load a tokenizer for a BERT-like model
tokenizer = AutoTokenizer.from_pretrained("bert-base-uncased")
def count_tokens(text):
    return len(tokenizer.encode(text))

text_splitter = CharacterTextSplitter(
    # separator="\n",
    chunk_size=400,
    chunk_overlap=50,
    # length_function=len,
    length_function=count_tokens,
    # is_separator_regex=False,
)
texts = text_splitter.create_documents([document])

import os
import json

directory = "./sections"
for file in os.listdir(directory):
    filename = os.fsdecode(file)
    if filename.endswith(".json"):
        with open(os.path.join(directory, filename)) as f:
            jsn = json.loads(f.read())
            section = jsn["section"]
            texts = text_splitter.create_documents([section])
            chunks = []
            for i,t in enumerate(texts):
                # replace newlines with spaces this can help keep word boundires
                chunks.append(t.page_content.replace("\n", " "))
            jsn["chunks"] = chunks

            with open(f'./chunks/{jsn["id"]}.json', "w") as wf:
                wf.write(json.dumps(jsn))
