==> ENGE817/Report/3_mimic_III_preprocess.ipynb <==
# Mimic II Dataset

https://physionet.org/content/mimiciii/1.4/

DOI (version 1.4):
https://doi.org/10.13026/C2HM2Q

DOI (latest version):
https://doi.org/10.13026/jbmn-w042

Topics:
critical care mimic electronic health records

Project Website:
https://mimic.physionet.org


## Data Description

MIMIC-III is a relational database consisting of 26 tables. Tables are linked by identifiers which usually have the suffix ‘ID’. For example, SUBJECT_ID refers to a unique patient, HADM_ID refers to a unique admission to the hospital, and ICUSTAY_ID refers to a unique admission to an intensive care unit.

Charted events such as notes, laboratory tests, and fluid balance are stored in a series of ‘events’ tables. For example the OUTPUTEVENTS table contains all measurements related to output for a given patient, while the LABEVENTS table contains laboratory test results for a patient.

Tables prefixed with ‘D_’ are dictionary tables and provide definitions for identifiers. For example, every row of CHARTEVENTS is associated with a single ITEMID which represents the concept measured, but it does not contain the actual name of the measurement. By joining CHARTEVENTS and D_ITEMS on ITEMID, it is possible to identify the concept represented by a given ITEMID.

Developing the MIMIC data model involved balancing simplicity of interpretation against closeness to ground truth. As such, the model is a reflection of underlying data sources, modified over iterations of the MIMIC database in response to user feedback. Care has been taken to avoid making assumptions about the underlying data when carrying out transformations, so MIMIC-III closely represents the raw hospital data.

Broadly speaking, five tables are used to define and track patient stays: ADMISSIONS; PATIENTS; ICUSTAYS; SERVICES; and TRANSFERS. Another five tables are dictionaries for cross-referencing codes against their respective definitions: D_CPT; D_ICD_DIAGNOSES; D_ICD_PROCEDURES; D_ITEMS; and D_LABITEMS. The remaining tables contain data associated with patient care, such as physiological measurements, caregiver observations, and billing information.

In some cases it would be possible to merge tables—for example, the D_ICD_PROCEDURES and CPTEVENTS tables both contain detail relating to procedures and could be combined—but our approach is to keep the tables independent for clarity, since the data sources are significantly different. Rather than combining the tables within MIMIC data model, we suggest researchers develop database views and transforms as appropriate.

## Usage Notes

MIMIC-III is provided as a collection of comma separated value (CSV) files, along with scripts to help with importing the data into database systems including PostreSQL, MySQL, and MonetDB. As the database contains detailed information regarding the clinical care of patients, it must be treated with appropriate care and respect. Researchers are required to formally request access via a process documented on the MIMIC website. There are two key steps that must be completed before access is granted:

- the researcher must complete a recognized course in protecting human research participants that includes Health Insurance Portability and Accountability Act (HIPAA) requirements.
- the researcher must sign a data use agreement, which outlines appropriate data usage and security standards, and forbids efforts to identify individual patients.

Approval requires at least a week. Once an application has been approved the researcher will receive emails containing instructions for downloading the database from PhysioNetWorks, a restricted access component of PhysioNet.

%pip install duckdb pandas

import duckdb
import pandas as pd

## Import MIMIC-III Data Into DuckDB

if False:
    conn = duckdb.connect('mimiciii.db', read_only=False)
    try:
        df = conn.execute("""
            -- To get unique patients
            CREATE TABLE IF NOT EXISTS PATIENTS AS
                FROM read_csv_auto(
                            'datasets/physionet.org/files/mimiciii/1.4/PATIENTS.csv.gz',
                            header=True,
                            union_by_name=true,
                            files_to_sniff=-1,
                            filename=true
                            );

            -- To get a single visit and dianosis. Without this it's hard
            -- to limit a patient diagnosis to an event
            CREATE TABLE IF NOT EXISTS ADMISSIONS AS
                FROM read_csv_auto(
                            'datasets/physionet.org/files/mimiciii/1.4/ADMISSIONS.csv.gz',
                            header=True,
                            union_by_name=true,
                            files_to_sniff=-1,
                            filename=true
                            );

            -- Not 100% I need this one just yet
            CREATE TABLE IF NOT EXISTS CHARTEVENTS AS
                FROM read_csv_auto(
                            'datasets/physionet.org/files/mimiciii/1.4/CHARTEVENTS.csv.gz',
                            header=True,
                            union_by_name=true,
                            files_to_sniff=-1,
                            filename=true,
                            sample_size=-1
                            );

            -- ICD9_CODE (used to get diagnosis id)
            CREATE TABLE IF NOT EXISTS DIAGNOSES_ICD AS
                FROM read_csv_auto(
                            'datasets/physionet.org/files/mimiciii/1.4/DIAGNOSES_ICD.csv.gz',
                            header=True,
                            union_by_name=true,
                            files_to_sniff=-1,
                            filename=true,
                            sample_size=-1
                            );

            -- ICD9_CODE (diagnosis text)
            CREATE TABLE IF NOT EXISTS D_ICD_DIAGNOSES AS
                FROM read_csv_auto(
                            'datasets/physionet.org/files/mimiciii/1.4/D_ICD_DIAGNOSES.csv.gz',
                            header=True,
                            union_by_name=true,
                            files_to_sniff=-1,
                            filename=true,
                            sample_size=-1
                            );

            -- to get the patients chart.
            CREATE TABLE IF NOT EXISTS NOTEEVENTS AS
                FROM read_csv_auto(
                            'datasets/physionet.org/files/mimiciii/1.4/NOTEEVENTS.csv.gz',
                            header=True,
                            union_by_name=true,
                            files_to_sniff=-1,
                            filename=true,
                            sample_size=-1
                            );
        """).df()
        print(df)

        # conn.execute("COPY mytable TO 'output.csv' (HEADER, DELIMITER ',')")

    except Exception as ex:
        print(ex)
    finally:
        conn.close()


## Classes in the Vector Database as Integers

from tables import classes
print(classes[0])
print(classes[10])
print(classes[20])
print(classes[30])
print(classes[34])

## Create Patient Record Data Linked to Diagnosis

| Category | Description |
|----------|--------------|
| Discharge Summary	|A comprehensive summary of a patient's hospital stay, including diagnoses, treatments, and follow-up plans.|
|Nursing Notes	|Documentation by nursing staff detailing patient care, observations, and any changes in condition.|
|Physician Notes	|Notes written by physicians that include assessments, treatment plans, and clinical observations.|
|Radiology Reports	|Reports generated from imaging studies, detailing findings and interpretations of the images.|
|Consult Notes	|Notes from specialists who provide additional insights or recommendations regarding a patient's care.|
|Procedure Notes	|Documentation of any procedures performed, including details about the procedure and patient response.|
|Medication Orders	|Records of medications prescribed, including dosages and administration instructions.|
|Progress Notes	|Ongoing documentation of a patient's progress, including updates on treatment and response to care.|

# "Depressive Disorders" == "Depressive disorder NEC"
#    "Anxiety Disorders" == "Anxiety state NOS"
conn = duckdb.connect('mimiciii.db', read_only=True)
result = conn.sql("""
    SELECT
        CASE
            WHEN dd.short_title = 'Depressive disorder NEC' THEN 10
            -- WHEN dd.short_title = 'Anxiety state NOS' THEN 11
        END AS class_number,
        count(*)
    FROM patients p
    JOIN diagnoses_icd d ON p.subject_id = d.subject_id
    JOIN d_icd_diagnoses dd ON d.icd9_code = dd.icd9_code
    JOIN (
        SELECT t2.subject_id,
            ROW_NUMBER() OVER(PARTITION BY t2.subject_id) as rn
        FROM noteevents t2
    ) AS t2 ON (p.subject_id = t2.subject_id AND t2.rn = 1)
    WHERE short_title = 'Depressive disorder NEC'
       -- OR short_title = 'Anxiety state NOS'
    GROUP BY class_number
""").fetchall()
print(result)
conn.close()


## Negitive Examples

Create a list of Subject IDs that do have a diagnosis of Depression

# select distinct(short_title) from d_icd_diagnoses dd where dd.short_title like 'Schizophrenia%'; == 7
# select distinct(short_title) from d_icd_diagnoses dd where dd.short_title like 'Catatonia%';     == 8
# select distinct(short_title) from d_icd_diagnoses dd where dd.short_title like 'Bipolar%';       == 9
# select distinct(short_title) from d_icd_diagnoses dd where dd.short_title like 'Depressive%';    == 10  / 311, 296.20    <----
# select distinct(short_title) from d_icd_diagnoses dd where dd.short_title like 'Anxiety%';       == 11 / 30000, 29384
# select distinct(short_title) from d_icd_diagnoses dd where dd.short_title like 'Pulmonary%';      example

# total patients: 46520
# count with anxiety: 1454
# count with depression: 3067
# count with depression and anxiety: 4192
# count without depression or anxiety: 42328

conn = duckdb.connect('mimiciii.db', read_only=True)
result = conn.sql("""
    COPY (
        SELECT
            pp.subject_id,
            -- GROUP_CONCAT(nee.text) AS chart_text,
            CASE
                WHEN TRY_CAST(ddi.icd9_code AS int32) =   311 THEN 10
                WHEN TRY_CAST(ddi.icd9_code AS int32) = 29620 THEN 10
                ELSE 0
            END AS class_number
        FROM patients pp
        JOIN diagnoses_icd di ON pp.subject_id = di.subject_id
        JOIN d_icd_diagnoses ddi ON di.icd9_code = ddi.icd9_code
        JOIN noteevents ne ON pp.subject_id = ne.subject_id
        WHERE pp.subject_id NOT IN
        (
            SELECT p.subject_id
            FROM patients p
            JOIN diagnoses_icd d ON p.subject_id = d.subject_id
            WHERE (
                -- depression
                TRY_CAST(d.icd9_code as int32) = 311
                OR TRY_CAST(d.icd9_code as int32) = 29620
            )
            GROUP BY p.subject_id
        )
        AND (
           -- ne.category = 'Discharge summary'
           ne.category = 'Nursing'
           OR ne.category = 'Physician'
        )
        GROUP BY pp.subject_id, class_number
        ORDER BY class_number DESC
        LIMIT 100
    ) TO 'datasets/negitive.csv' (FORMAT csv, DELIMITER '|', HEADER, QUOTE '"');
""")
#print(len(result))
#display(result)
conn.close()

## Positive Examples

Create a list of subject id, class, and chart text for patients that had a depression diagnosis 

# "Depressive Disorders" == "Depressive disorder NEC"
#    "Anxiety Disorders" == "Anxiety state NOS"
conn = duckdb.connect('mimiciii.db', read_only=True)
conn.sql("""
    COPY (
        SELECT
            p.subject_id,
            CASE
                WHEN TRY_CAST(ddi.icd9_code AS int32) =   311 THEN 10
                WHEN TRY_CAST(ddi.icd9_code AS int32) = 29620 THEN 10
                ELSE 0
            END AS class_number,
            GROUP_CONCAT(ne.text) AS chart_text
        FROM patients p
        JOIN diagnoses_icd d ON p.subject_id = d.subject_id
        JOIN d_icd_diagnoses ddi ON d.icd9_code = ddi.icd9_code
        JOIN noteevents ne ON p.subject_id = ne.subject_id
        WHERE (
            TRY_CAST(ddi.icd9_code AS int32) IN (311, 29620)
        )
        AND (
           -- ne.category = 'Discharge summary'
           ne.category = 'Nursing'
           OR ne.category = 'Physician'
        )
        GROUP BY p.subject_id, class_number
        LIMIT 100
    ) TO 'datasets/positive_w_text.csv' (FORMAT csv, DELIMITER '|', HEADER, QUOTE '"');
""") #.write_csv('datasets/text_classes.csv', sep="|", escapechar="~")
# display(len(result)) # ==> 4353
conn.close()

## Populate Negitive Examples

As an extra step, we need to grab the chart data for the negitive examples 

df = pd.read_csv("datasets/negitive.csv", delimiter='|')
conn = duckdb.connect('mimiciii.db', read_only=True)

def get_patient_notes(row):
    subject_id = row['SUBJECT_ID']
    qresult = conn.sql(f"""
        SELECT
            p.subject_id as id,
            0 as class_number,
            GROUP_CONCAT(ne.text) AS chart_text
        FROM patients p
        JOIN diagnoses_icd d ON p.subject_id = d.subject_id
        JOIN d_icd_diagnoses ddi ON d.icd9_code = ddi.icd9_code
        JOIN noteevents ne ON p.subject_id = ne.subject_id
        WHERE p.subject_id = {subject_id}
        AND (
            -- ne.category = 'Discharge summary'
            ne.category = 'Nursing'
            OR ne.category = 'Physician'
        )
        GROUP BY p.subject_id, class_number
    """).fetchall()
    for record in qresult:
        row["chart_text"] = record[2]
    return row

result = df.apply(get_patient_notes, axis=1)
conn.close()
# display(result)
result.to_csv("datasets/negitive_w_text.csv", sep='|', quotechar='"', index=False, encoding='utf-8')
