==> ENGE817/Report/5_results.ipynb <==
import pandas as pd
import numpy as np
from tables import schema_dicts

result_set = "results/1800_nurse_phy_cosine"

$ a = \frac{TP+TN}{TP+TN+FP+FN} $

$ p = \frac{TP}{TP+FP} $

$ r = \frac{TP}{TP+FN} $

$ f_1 = 2 \cdot \frac{p \cdot r}{p + r} $

for i in range(4):
    print(f'======== {schema_dicts[i]["name"]} ==========')
    df = pd.read_csv(f"{result_set}/{i}_20_1800.csv", index_col=0)
    # just to make it easier to read
    del df['predict']

    def array_from_string(x):
        return np.fromstring(x.strip('[]'), sep=' ')

    def tp(orig, top5):
        return 1 if orig == 10 and 10 in top5 else 0

    def tn(orig, top5):
        return 1 if orig != 10 and 10 not in top5 else 0

    def fp(orig, top5):
        return 1 if orig == 0 and 10 in top5 else 0

    def fn(orig, top5):
        return 1 if orig == 10 and 10 not in top5 else 0

    df['top5'] = df['top5'].map(array_from_string)
    df['tp'] = df.apply(lambda x: tp(orig=x['original'], top5=x['top5']), axis=1)
    df['tn'] = df.apply(lambda x: tn(orig=x['original'], top5=x['top5']), axis=1)
    df['fp'] = df.apply(lambda x: fp(orig=x['original'], top5=x['top5']), axis=1)
    df['fn'] = df.apply(lambda x: fn(orig=x['original'], top5=x['top5']), axis=1)

    display(df)

    TP = df['tp'].sum()
    TN = df['tn'].sum()
    FP = df['fp'].sum()
    FN = df['fn'].sum()

    print(f"TP    {TP:.2f}")
    print(f"TN    {TN:.2f}")
    print(f"FP    {FP:.2f}")
    print(f"FN    {FN:.2f}")
    print("----------------")
    print(TP+TN+FP+FN)

    accuracy = (TP + TN) / (TP + TN + FP + FN)
    precision = (TP / (TP+FP))
    recall = (TP / (TP+FN))
    print(f"A    {accuracy:.2f}")
    print(f"P    {precision:.2f}")
    print(f"R    {recall:.2f}")

    f1 = 2 * ((precision*recall) / (precision+recall))
    print("----------------")
    print(f"F1    {f1:.2f}")
    print("----------------")

    # print(f"Correct top 5    {df['inTop5'].sum() / len(df)}")
    # print(f"Correct category {df['correct'].sum() / len(df)}")

    column_counts = df["original"].value_counts(dropna=False)
    # print(column_counts)
    print(f'==========================================')
    print("")
